requirements.txt
Django==5.0.1
django-crispy-forms==2.1
crispy-bootstrap4
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-taggit==5.0.1
djangorestframework==3.14.0
social-auth-app-django==5.4.0
